friendlist-updated
- modules added for dom query elements and fetch function api.
- used one common module.
