+++ assignment-13.htm
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="styles/css/friendlist.css">
 
     <!--script reference added-->
-    <script src="js/friendlist.js" defer></script>
+    <script src="js/friendlist.js" type="module"></script>
   </head>
   <body>
     <!--parent_container-->
