+++ js/domconstants.js
@@ -1,0 +1,5 @@
+// selecting friendList container
+const friendListSpace = document.querySelector(".friendList");
+const fragment = document.createDocumentFragment();
+
+export { fragment, friendListSpace };
