+++ styles/css/friendlist.css
@@ -25,7 +25,6 @@ body {
   background-color: #3b1c67;
   box-sizing: border-box;
   color: #ffff;
-  padding-top: 30px;
   /* padding-bottom: 30px; */
   padding: 30px 0 30px 35px;
   /* padding-left: 35px;  */
